moved the ConnectionManager
